requirements.txt
torch
transformers
transformer_lens
datasets
numpy
pandas
matplotlib
seaborn
tqdm
einops
scipy
pyyaml
